Rubric Section 2: Meets
Updated social media links to point to external websites.

Did not have to delete social media icons because a link was provided
for each.

Edited the navigation link for Home to navigate to the index.html.

Edited the navigation link for Experience to navigate to
experience.html and uploaded a duplicate of index.html as
experience.html to use as a template.